+++ CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ## 0.0.2 (November 7, 2018)
 
-+ `F8` run selection and delete
++ `F7` run selection and delete
 + error log
 + workspace log
 + log when run scripts
 + highlight setting for `%%` `% [...]`
++ script queue
 
 + fix: clear before run
 + fix: `F9` support `...`breaking line
 + fix: support `import`
++ fix: `F5` support `\+a\+b\c.m`=>`a.b.c`
 
 ## 0.0.1 (November 1, 2018)